Simple scala plugin for gerrit
This is a very simple HelloWorld plugin trying to demonstrate that
we can react correctly to a REST call when used by Gerrit.
This is simply answering to $GERRITURL/plugins/generic-scala-plugin/
writing [generic scala plugin hello world] 1-2-3
The 1-2-3 part is explicitly using scala runtime via List to be sure
we are using correctly it from gerrit.
The sbt or bazel packager must embed needed runtime in final .jar